Python update (#470)
Recent numpy / python updates broke several tests.

Fixed path in test.
Fixed deprecated numpy decorator warning. See
https://numba.readthedocs.io/en/stable/reference/deprecation.html#deprecation-of-object-mode-fall-back-behaviour-when-using-jit.
Removed deprecated np.int type.

+++ pftools/python/parflow/tools/builders.py
@@ -1378,7 +1378,7 @@ class CLMImporter:
             array = vegm_data[:, :, i]
             if token == 'Color':
                 # This one needs to be an integer
-                array = array.astype(np.int)
+                array = array.astype(np.int32)
 
             if array.min() == array.max():
                 # All the values are the same
